# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/posts/images": {
      "post": {
        "operationId": "PostsController_createImagePost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImagePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/api/v1/posts/{postId}": {
      "get": {
        "operationId": "PostsController_findOne",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyPostDto"
                }
              }
            }
          },
          "404": {
            "description": "Post does not exist or is not public"
          }
        },
        "summary": "View one public published post",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/v1/playlists": {
      "post": {
        "operationId": "PlaylistsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "playlists"
        ]
      }
    },
...
    "/api/v1/playlists/{playlistId}": {
      "get": {
        "operationId": "PlaylistsController_public",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "playlists"
        ]
      },
      "patch": {
        "operationId": "PlaylistsController_update",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "playlists"
        ]
      },
      "delete": {
        "operationId": "PlaylistsController_delete",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/v1/playlists/{playlistId}/manage": {
      "get": {
        "operationId": "PlaylistsController_manage",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/v1/playlists/{playlistId}/items": {
      "post": {
        "operationId": "PlaylistsController_add",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPlaylistItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/v1/playlists/{playlistId}/items/order": {
      "put": {
        "operationId": "PlaylistsController_reorder",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderPlaylistItemsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "playlists"
        ]
      }
    },
    "/api/v1/playlists/{playlistId}/items/{itemId}": {
      "delete": {
        "operationId": "PlaylistsController_remove",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "playlists"
        ]
      }
    }
  },
  "info": {
    "title": "Stream Platform API",
    "description": "Public and authenticated media platform operations",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatorSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "handle",
          "displayName"
        ]
      },
...
      "MediaSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "durationMs": {
            "type": "number",
            "nullable": true
          },
          "posterUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "url",
          "mimeType",
          "width",
          "height"
        ]
      },
...
      "CreateImagePostDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "caption": {
            "type": "string",
            "maxLength": 2200
          }
        },
        "required": [
          "assetId",
          "caption"
        ]
      },
      "LegacyPostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "format": {
            "type": "string",
            "enum": [
              "IMAGE",
              "SHORT_VIDEO",
              "LONG_VIDEO",
              "COMMUNITY_TEXT",
              "COMMUNITY_VIDEO",
              "COMMUNITY_LINK"
            ]
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "caption": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string"
          },
          "likeCount": {
            "type": "number"
          },
          "commentCount": {
            "type": "number"
          },
          "author": {
            "$ref": "#/components/schemas/CreatorSummaryDto"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "format",
          "caption",
          "publishedAt",
          "likeCount",
          "commentCount",
          "author",
          "media"
        ]
      },
...
      "CreatePlaylistDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "UNLISTED",
              "PUBLIC"
            ],
            "default": "PRIVATE"
          }
        },
        "required": [
          "title"
        ]
      },
      "PlaylistTargetDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "HOME_VIDEO",
              "SERIES",
              "SERIES_EPISODE",
              "SHORTFORM",
              "COMMUNITY_POST"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "type",
          "id"
        ]
      },
      "PlaylistItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "number"
          },
          "available": {
            "type": "boolean"
          },
          "target": {
            "$ref": "#/components/schemas/PlaylistTargetDto"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "href": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "position",
          "available",
          "target"
        ]
      },
      "PlaylistDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "UNLISTED",
              "PUBLIC"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistItemDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "visibility",
          "items"
        ]
      },
...
      "UpdatePlaylistDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "UNLISTED",
              "PUBLIC"
            ]
          }
        }
      },
      "AddPlaylistItemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "HOME_VIDEO",
              "SERIES",
              "SERIES_EPISODE",
              "SHORTFORM",
              "COMMUNITY_POST"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "type",
          "id"
        ]
      },
      "ReorderPlaylistItemsDto": {
        "type": "object",
        "properties": {
          "itemIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "itemIds"
        ]
      }
    }
  }
}